Process Results Scripts
Reads in the SALUS seasonal csvs with pandas library.
Calculates the seasonal average, min, and max of all the variables over the 38 years. (Results/Summary_1979_2016/).
The last value in 2016 for each of the variables is written out to a file (Results/End_of_2016/).
The script also writes out the yearly GWAD to a separate file (Results/Yearly_1979_2016/).

The Missing Soils script calculated the percentage of area of missing data due to missing SSURGO Mukey characteristics by state and the whole Midwest