Working HPC scripts
The directories with Final_HPC indicate the HPC scripts that are being used.
The final method of processing the Results was to create a Midwest Baseline raster with MUKEY and weather ID as attributes. Then to merge these attributes with the SALUS results to get a raster that you can classify using any attribute.