Updated method for creating baseline raster
The previous script did not mosaic the raster pieces (attribute tables) correctly. Therefore the weather ID values were messed up. Now I am trying to overlay the state soils with the state weather polygon on the HPC (the desktop runs out of memory).
Then the unique combination may need to be rerun with ExtractSoilWx.R for the SALUS runs.
Then the state baseline rasters need to be copied from the HPC and combined together using the mosaic tool on ArcGIS.